Change NumberOfAttemptsLeft to unsigned integer & format

+++ Source/Pioneer/PioneerCartridge.h
@@ -1,19 +1,16 @@
-// Fill out your copyright notice in the Description page of Project Settings.
-
 #pragma once
 
 #include "CoreMinimal.h"
 #include "Console/Cartridge.h"
 #include "PioneerCartridge.generated.h"
 
-
-UCLASS(ClassGroup=(Custom), meta=(BlueprintSpawnableComponent))
+UCLASS(ClassGroup = (Custom), meta = (BlueprintSpawnableComponent))
 class PIONEER_API UPioneerCartridge : public UCartridge
 {
 
 	GENERATED_BODY()
 
-	public:
+public:
 	void SetupGuessingGame();
 	void HandleInvalidGuess(FString message);
 	void ResetAttempts();
@@ -22,11 +19,11 @@ class PIONEER_API UPioneerCartridge : public UCartridge
 	void HandleGuess(FString Guess);
 
 	virtual void BeginPlay() override;
-	virtual void OnInput(const FString& Input) override;
+	virtual void OnInput(const FString &Input) override;
 
 	FString ValidateGuess(FString Response);
 
-	private:
+private:
 	FString Secret;
-	int32 NumberOfAttemptsLeft;
+	uint8 NumberOfAttemptsLeft;
 };
